Factor out to composable
